requirements.txt
Django==1.9.8
django-ckeditor==5.0.3
django-facebook==6.0.3
django-markdown==0.8.4
facebook-sdk==2.0.0
Markdown==2.6.6
mysqlclient==1.3.7
requests==2.11.0
Unidecode==0.4.19
